PIM-3873: move repositories
