+++ spec/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/ProductMassActionRepositorySpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\CatalogBundle\Doctrine\ORM;
+namespace spec\Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\EntityManager;
 use PhpSpec\ObjectBehavior;
