+++ spec/Pim/Bundle/DataGridBundle/Datasource/DatasourceSpec.php
@@ -5,14 +5,14 @@ namespace spec\Pim\Bundle\DataGridBundle\Datasource;
 use Doctrine\Common\Persistence\ObjectManager;
 use Oro\Bundle\DataGridBundle\Datagrid\DatagridInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductMassActionRepository;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductRepository;
+use Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductRepository;
+use Pim\Bundle\CatalogBundle\Repository\ProductMassActionRepositoryInterface;
 use Pim\Bundle\DataGridBundle\Datasource\ResultRecord\HydratorInterface;
 use Pim\Bundle\CatalogBundle\Query\ProductQueryBuilderFactoryInterface;
 
 class DatasourceSpec extends ObjectBehavior
 {
-    function let(ObjectManager $manager, HydratorInterface $hydrator, ProductMassActionRepository $massActionRepo, ProductQueryBuilderFactoryInterface $factory)
+    function let(ObjectManager $manager, HydratorInterface $hydrator, ProductMassActionRepositoryInterface $massActionRepo, ProductQueryBuilderFactoryInterface $factory)
     {
         $this->beConstructedWith($manager, $hydrator, $massActionRepo, $factory);
     }
