+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/AttributeGroupRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Entity\Repository;
+namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\AbstractQuery;
 use Pim\Bundle\CatalogBundle\Doctrine\ReferableEntityRepository;
@@ -13,8 +13,6 @@ use Pim\Bundle\CatalogBundle\Repository\AttributeGroupRepositoryInterface;
  * @author    Gildas Quemener <gildas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated will be moved to Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository in 1.4
  */
 class AttributeGroupRepository extends ReferableEntityRepository implements AttributeGroupRepositoryInterface
 {
