+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/CompletenessRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Doctrine\ORM;
+namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\EntityManager;
 use Pim\Bundle\CatalogBundle\Repository\CompletenessRepositoryInterface;
@@ -11,8 +11,6 @@ use Pim\Bundle\CatalogBundle\Repository\CompletenessRepositoryInterface;
  * @author    Benoit Jacquemont <benoit@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated will be moved to Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository in 1.4
  */
 class CompletenessRepository implements CompletenessRepositoryInterface
 {
