+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/CurrencyRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Entity\Repository;
+namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 
 use Pim\Bundle\CatalogBundle\Doctrine\ReferableEntityRepository;
 use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
@@ -11,8 +11,6 @@ use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated will be moved to Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository in 1.4
  */
 class CurrencyRepository extends ReferableEntityRepository implements CurrencyRepositoryInterface
 {
