+++ src/Pim/Bundle/CatalogBundle/Resources/config/repositories.yml
@@ -1,15 +1,15 @@
 parameters:
-    pim_catalog.repository.association_type.class: Pim\Bundle\CatalogBundle\Entity\Repository\AssociationTypeRepository
-    pim_catalog.repository.attribute_group.class:  Pim\Bundle\CatalogBundle\Entity\Repository\AttributeGroupRepository
-    pim_catalog.repository.attribute_option.class: Pim\Bundle\CatalogBundle\Entity\Repository\AttributeOptionRepository
-    pim_catalog.repository.attribute.class:        Pim\Bundle\CatalogBundle\Entity\Repository\AttributeRepository
-    pim_catalog.repository.category.class:         Pim\Bundle\CatalogBundle\Entity\Repository\CategoryRepository
-    pim_catalog.repository.channel.class:          Pim\Bundle\CatalogBundle\Entity\Repository\ChannelRepository
-    pim_catalog.repository.currency.class:         Pim\Bundle\CatalogBundle\Entity\Repository\CurrencyRepository
-    pim_catalog.repository.family.class:           Pim\Bundle\CatalogBundle\Entity\Repository\FamilyRepository
-    pim_catalog.repository.group.class:            Pim\Bundle\CatalogBundle\Entity\Repository\GroupRepository
-    pim_catalog.repository.group_type.class:       Pim\Bundle\CatalogBundle\Entity\Repository\GroupTypeRepository
-    pim_catalog.repository.locale.class:           Pim\Bundle\CatalogBundle\Entity\Repository\LocaleRepository
+    pim_catalog.repository.association_type.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AssociationTypeRepository
+    pim_catalog.repository.attribute_group.class:  Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AttributeGroupRepository
+    pim_catalog.repository.attribute_option.class: Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AttributeOptionRepository
+    pim_catalog.repository.attribute.class:        Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AttributeRepository
+    pim_catalog.repository.category.class:         Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\CategoryRepository
+    pim_catalog.repository.channel.class:          Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ChannelRepository
+    pim_catalog.repository.currency.class:         Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\CurrencyRepository
+    pim_catalog.repository.family.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\FamilyRepository
+    pim_catalog.repository.group.class:            Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\GroupRepository
+    pim_catalog.repository.group_type.class:       Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\GroupTypeRepository
+    pim_catalog.repository.locale.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\LocaleRepository
     pim_catalog.repository.product_template.class: Pim\Bundle\CatalogBundle\Entity\Repository\ProductTemplateRepository
     pim_catalog.repository.product_value.class:    Doctrine\ORM\EntityRepository
