+++ src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/orm.yml
@@ -1,9 +1,9 @@
 parameters:
-    pim_catalog.repository.association.class:                        Pim\Bundle\CatalogBundle\Doctrine\ORM\AssociationRepository
-    pim_catalog.repository.product.class:                            Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductRepository
-    pim_catalog.repository.completeness.class:                       Pim\Bundle\CatalogBundle\Doctrine\ORM\CompletenessRepository
-    pim_catalog.repository.product_mass_action.class:                Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductMassActionRepository
-    pim_catalog.repository.product_category.class:                   Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductCategoryRepository
+    pim_catalog.repository.association.class:                        Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AssociationRepository
+    pim_catalog.repository.product.class:                            Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductRepository
+    pim_catalog.repository.completeness.class:                       Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\CompletenessRepository
+    pim_catalog.repository.product_mass_action.class:                Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductMassActionRepository
+    pim_catalog.repository.product_category.class:                   Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductCategoryRepository
     pim_catalog.doctrine.completeness_generator.class:               Pim\Bundle\CatalogBundle\Doctrine\ORM\CompletenessGenerator
     pim_catalog.event_subscriber.orm.inject_product_reference.class: Pim\Bundle\CatalogBundle\EventSubscriber\ORM\InjectProductReferenceSubscriber
     pim_catalog.event_subscriber.metric_base_values.class:           Pim\Bundle\CatalogBundle\EventSubscriber\ORM\MetricBaseValuesSubscriber
