+++ src/Pim/Bundle/CatalogBundle/Validator/Constraints/VariantGroupValuesValidator.php
@@ -2,8 +2,8 @@
 
 namespace Pim\Bundle\CatalogBundle\Validator\Constraints;
 
-use Pim\Bundle\CatalogBundle\Entity\Repository\AttributeRepository;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
+use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 
@@ -17,9 +17,9 @@ use Symfony\Component\Validator\ConstraintValidator;
 class VariantGroupValuesValidator extends ConstraintValidator
 {
     /**
-     * @param AttributeRepository $attributeRepository
+     * @param AttributeRepositoryInterface $attributeRepository
      */
-    public function __construct(AttributeRepository $attributeRepository)
+    public function __construct(AttributeRepositoryInterface $attributeRepository)
     {
         $this->attributeRepository = $attributeRepository;
     }
